+++ scraper.rb
@@ -8,6 +8,9 @@ require 'scraped'
 require 'scraperwiki'
 require 'csv'
 require 'combine_popolo_memberships'
+require 'require_all'
+
+require_rel 'lib'
 
 # require 'open-uri/cached'
 # OpenURI::Cache.cache_path = '.cache'
@@ -16,98 +19,6 @@ require 'scraped_page_archive/open-uri'
 class String
   def tidy
     gsub(/[[:space:]]+/, ' ').strip
-  end
-end
-
-class CurrentMembersPage < Scraped::HTML
-  field :member_urls do
-    noko.css('.list__row').map do |entry|
-      URI.join(url, entry.css('a.theme__link/@href').text).to_s
-    end
-  end
-end
-
-class MembershipRow < Scraped::HTML
-  field :area do
-    td[0].text.tidy
-  end
-
-  field :party do
-    td[1].text.tidy
-  end
-
-  field :start_date do
-    '%d-%02d-%02d' % td[2].text.tidy.split('/').reverse.map(&:to_i)
-  end
-
-  field :end_date do
-    return unless td[3]
-    ed = td[3].text.tidy
-    return if ed.empty?
-    '%d-%02d-%02d' % ed.split('/').reverse.map(&:to_i)
-  end
-
-  private
-
-  def td
-    noko.css('td')
-  end
-end
-
-class CurrentMemberPage < Scraped::HTML
-  field :id do
-    url.to_s.split("/")[-2]
-  end
-
-  field :name do
-    raw_name.sub(/^Dr /,'').tidy
-  end
-
-  field :sort_name do
-    noko.css('title').text.split(' - ').first.tidy
-  end
-
-  # TODO use absolute URL decorator
-  field :photo do
-    raw = body.css('.document-panel__img img/@src').last.text
-    return if raw.to_s.empty?
-    URI.join(url, raw).to_s
-  end
-
-  field :email do
-    body.css('a.square-btn').attr('href').inner_text.gsub('mailto:','')
-  end
-
-  field :facebook do
-    body.css('div.related-links__item a[@href*="facebook"]/@href').text
-  end
-
-  field :twitter do
-    body.css('div.related-links__item a[@href*="twitter"]/@href').text
-  end
-
-  field :source do
-    url.to_s
-  end
-
-  field :honorific_prefix do
-    'Dr' if raw_name.start_with? 'Dr '
-  end
-
-  field :memberships do
-    noko.css('.body-text').xpath('//table[.//th[.="Party"]]').first.css('tr').map do |tr|
-      MembershipRow.new(response: response, noko: tr).to_h
-    end
-  end
-
-  private
-
-  def body
-    noko.css('div.koru-side-holder')
-  end
-
-  def raw_name
-    body.css("div[role='main'] h1").text.sub(/^(Rt )?Hon /,'').tidy
   end
 end
